Division de las funciones de accesder a archivos de las funciones de codigo
