+++ src/code/sede_code.py
@@ -12,59 +12,9 @@ Dependencias:
 - src.code.constantes.JSON_SEDE: Ruta del archivo JSON que contiene la información de las sedes.
 """
 
-import json
-import os
-from src.code.constantes import JSON_SEDE
+from tkinter import messagebox
 
-def agregar_sede_archivo(sede_id, nombre, ubicacion, estado, telefono):
-    """
-    Agrega una nueva sede al archivo JSON de sedes.
-
-    Parámetros:
-    - sede_id (str): ID de la sede.
-    - nombre (str): Nombre de la sede.
-    - ubicacion (str): Ubicación de la sede.
-    - estado (str): Estado de la sede.
-    - telefono (str): Número de teléfono de la sede.
-    """
-    nueva_sede = {
-        "id": sede_id,
-        "nombre": nombre,
-        "ubicacion": ubicacion,
-        "estado": estado,
-        "telefono": telefono
-    }
-
-    archivo_sedes = os.path.join(os.path.dirname(__file__), "..", "db", JSON_SEDE)
-    if os.path.exists(archivo_sedes):
-        with open(archivo_sedes, "r") as file:
-            data = json.load(file)
-    else:
-        data = {
-            "Sedes": []
-        }
-
-    data["Sedes"].append(nueva_sede)
-
-    with open(archivo_sedes, "w") as file:
-        json.dump(data, file, indent=4)
-
-
-def obtener_sedes():
-    """
-    Obtiene la lista de sedes del archivo JSON de sedes.
-
-    Retorna:
-    - list: Lista de sedes.
-    """
-    archivo_sedes = os.path.join(os.path.dirname(__file__), "..", "db", JSON_SEDE)
-
-    if os.path.exists(archivo_sedes):
-        with open(archivo_sedes, "r") as file:
-            data = json.load(file)
-            return data.get("Sedes", [])
-    else:
-        return []
+from src.code.storage.sede_storage import obtener_sedes
 
 
 def obtener_sedes_activas():
@@ -80,3 +30,54 @@ def obtener_sedes_activas():
         if sede["estado"] == "Activo":
             sedes_activas.append(sede["nombre"])
     return sedes_activas
+def validar_nombre_sede(nombre_sede):
+    sedes = obtener_sedes_activas()
+
+    for sede in sedes:
+        if sede == nombre_sede:
+            return True
+    messagebox.showerror("Error de sede", "Nombre de sede incorrecto")
+    return False
+
+def validar_campos_form_sede( nombre, ubicacion, estado, telefono):
+    """
+    Valida los campos del formulario.
+
+    Parámetros:
+    - nombre: El nombre de la sede.
+    - ubicacion: La ubicación de la sede.
+    - estado: El estado de la sede.
+    - telefono: El número de teléfono de la sede.
+
+    Retorna:
+    - bool: True si los campos son válidos, False en caso contrario.
+    """
+    sedes = obtener_sedes()
+    nombrevalido = True
+    for sede in sedes:
+        if sede["nombre"].upper() == nombre.upper():
+            nombrevalido = False
+    if not nombrevalido:
+        messagebox.showerror("Error", "Ya existe un sede con ese nombre")
+        return False
+    if not nombre or not ubicacion or not estado or not telefono:
+        messagebox.showerror("Error", "Por favor complete todos los campos.")
+        return False
+
+    try:
+        telefono = int(telefono)
+    except ValueError:
+        messagebox.showerror("Error", "El número de teléfono debe ser numérico.")
+        return False
+
+    if len(str(telefono)) != 8:
+        messagebox.showerror("Error", "El número de teléfono debe tener 8 dígitos.")
+        return False
+    return True
+
+
+def limpiar_formulario_sede(app):
+    app.nombre_var.set("")
+    app.ubicacion_var.set("")
+    app.estado_var.set("")
+    app.telefono_var.set("")
