+++ src/db/funcionarios.json
@@ -19,7 +19,7 @@
             "id": "F-4444",
             "nombre": "Juan Gonzales",
             "telefono": "87654321",
-            "idcentro": "CA115"
+            "idcentro": "CE1105"
         }
     ]
 }
