--- src/sedes.json
@@ -1,11 +1,0 @@
-{
-    "Sedes": [
-        {
-            "id": "S-INL1LNR5BU3T",
-            "nombre": "Pacifico",
-            "ubicacion": "Puntarenas",
-            "estado": "Activo",
-            "telefono": "89898989"
-        }
-    ]
-}
